backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
catboost==1.2.8
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepseek==1.0.0
distro==1.9.0
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.1
narwhals==1.38.2
numpy==2.2.5
openai==1.77.0
openmeteo_requests==1.4.0
openmeteo_sdk==1.20.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.8
plotly==6.0.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.2
xgboost==3.0.0
